Pymol related function are defined only in Pymol
Pymol related functions which used the Pymol API were accessible
even out of the context of Pymol. This could have caused runtime errors
since the API was not available. These functions are now defined only
in the context of Pymol so the use of molndx as a module is safer.

The change is based on the fact that the "__name__" variable contains
"pymol" when the module is run in Pymol.