Solve the selection issue
The "index" selector was used instead of the "id" one in pymol. The
difference was not visible in most cases but could however causes the
selection to be shifted.

"index" means for "internal index number" and "id" for "external index
number".

The warning has been remove from the README. Issue #2 is fixed.

No unit test have been developed to check this issue yet. This will be
a thing to do.